Use trick to reduce cancellation error when computing Binomial and Poisson probabilities in log-space.
