+++ .editorconfig
@@ -3,3 +3,6 @@
 [*.cs]
 # bracing and indenting
 csharp_new_line_before_open_brace = none
+csharp_new_line_before_else = false
+csharp_new_line_before_catch = false
+csharp_new_line_before_finally = false
